+++ extensions/iceworks-style-helper/src/index.ts
@@ -1,8 +1,13 @@
 import * as vscode from 'vscode';
+import cssClassAutoCompete from './cssClassAutoCompete';
 import inlineStyleAutoComplete from './inlineStyleAutoComplete';
+import styleInfoViewer from './styleInfoViewer';
+
 
 function activate(context: vscode.ExtensionContext) {
+  cssClassAutoCompete(context);
   inlineStyleAutoComplete(context);
+  styleInfoViewer(context);
 };
 
 exports.activate = activate;