+++ packages/project-service/src/index.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import * as fsExtra from 'fs-extra';
 import { downloadAndGenerateProject } from '@iceworks/generate-project';
 import { IMaterialScaffold } from '@iceworks/material-utils';
-import { checkPathExists } from '@iceworks/common-service';
+import { checkPathExists, getDataFromSettingJson, CONFIGURATION_KEY_NPM_REGISTRY } from '@iceworks/common-service';
 import { readPackageJSON } from 'ice-npm-utils';
 import * as simpleGit from 'simple-git/promise';
 import * as path from 'path';
@@ -91,9 +91,10 @@ export async function createProject(data): Promise<string> {
   const projectDir: string = path.join(projectPath, projectName);
   const isProjectDirExists = await checkPathExists(projectDir);
   if (isProjectDirExists) {
-    throw new Error(`文件夹「${projectDir}」已存在，请重新输入项目名称。`)
+    throw new Error(`文件夹「${projectDir}」已存在，请重新输入应用名称。`)
   }
-  const { npm, registry, version } = scaffold.source;
+  const { npm, version } = scaffold.source;
+  const registry = getDataFromSettingJson(CONFIGURATION_KEY_NPM_REGISTRY);
   await downloadAndGenerateProject(projectDir, npm, version, registry);
   return projectDir;
 }
@@ -112,7 +113,7 @@ export async function CreateDEFProjectAndCloneRepository(DEFProjectField: IDEFPr
   const projectDir = path.join(projectPath, projectName);
   const isProjectDirExists = await checkPathExists(projectDir);
   if (isProjectDirExists) {
-    throw new Error(`文件夹「${projectDir}」已存在，请重新输入项目名称。`)
+    throw new Error(`文件夹「${projectDir}」已存在，请重新输入应用名称。`)
   }
   await createDEFProject(DEFProjectField);
   await cloneRepositoryToLocal(projectDir, group, project);
@@ -130,7 +131,7 @@ export async function createDEFProject(DEFProjectField: IDEFProjectField): Promi
 async function cloneRepositoryToLocal(projectDir, group, project): Promise<void> {
   const isProjectDirExists = await checkPathExists(projectDir);
   if (isProjectDirExists) {
-    throw new Error(`文件夹「${projectDir}」已存在，请重新输入项目名称。`)
+    throw new Error(`文件夹「${projectDir}」已存在，请重新输入应用名称。`)
   }
   const repoPath = `git@gitlab.alibaba-inc.com:${group}/${project}.git`;
   await simpleGit().clone(repoPath, projectDir)
@@ -183,10 +184,10 @@ function getGeneratorTaskStatus(taskId: number, clientToken: string): Promise<an
         if (status !== GeneratorTaskStatus.running && status !== GeneratorTaskStatus.Created) {
           clearInterval(interval);
           if (status === GeneratorTaskStatus.Failed) {
-            reject(new Error(`创建项目失败，任务 ID 是： ${taskId}.`))
+            reject(new Error(`创建 DEF 应用失败，任务 ID 是： ${taskId}.`))
           }
           if (status === GeneratorTaskStatus.Timeout) {
-            reject(new Error(`创建项目超时，任务 ID 是：${taskId}.`))
+            reject(new Error(`创建 DEF 应用超时，任务 ID 是：${taskId}.`))
           }
           if (status === GeneratorTaskStatus.Success) {
             resolve()