+++ extensions/iceworks-style-helper/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.2.0
+
+Add `class` name completions.
+
 ## 0.1.0
 
 Initial release