+++ extensions/iceworks/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 0.1.5
+
+- docs: project rename to application
+- docs: add homepage and repository 
+
+## 0.1.4
+
+- docs: iceworks & iceworks-material-help & iceworks-style-help
+- docs: quick start
+
 ## 0.1.3
 
 Add the following extensions: